split-files: Added instructions.py
Moved functions for writing the instructions file to a new file.

Issue: #22
